Сторонняя помощь, оружия стреляют, но связь пуль, игрока, оружий, пуллов не продумана
Много нюансов здесь. Первый- везде одни ивенты на перерасчёт и сет переменных. Второй- нельзя в классе брать переменные из SO даже если ты объявил префаб в этом классе!!! Это самая важная часть... Третий- много всего. Много бафов, много оружий, много проджектайлов. Куча пуллов. За всем надо следить... Я запутался жёстко. Сейчас думаю над организацией и взаимодействием объектов.

+++ Assets/ScriptableObjects/PlayerBuffs/PlayerBuffsSO.cs
@@ -19,7 +19,7 @@ public class PlayerBuffsSO : ScriptableObject
     public float PlayerBuffsShield = 3;
     public float PlayerBuffsSpeed = 5f;
 
-    public float PlayerBuffsPower = 25f;
+    public int PlayerBuffsPower = 25;
     public float PlayerBuffsProjectileSpeed = 20f;
     public float PlayerBuffsRrange = 11f;
     public float PlayerBuffsReload = -5f;
